feat: move autotasks folder
