+++ autotasks/aave-harvester/src/factories/AggregatorV3Interface.d.ts
@@ -121,9 +121,7 @@ export class AggregatorV3Interface extends BaseContract {
       }
     >;
 
-    latestRoundData(
-      overrides?: CallOverrides
-    ): Promise<
+    latestRoundData(overrides?: CallOverrides): Promise<
       [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
         roundId: BigNumber;
         answer: BigNumber;
@@ -153,9 +151,7 @@ export class AggregatorV3Interface extends BaseContract {
     }
   >;
 
-  latestRoundData(
-    overrides?: CallOverrides
-  ): Promise<
+  latestRoundData(overrides?: CallOverrides): Promise<
     [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
       roundId: BigNumber;
       answer: BigNumber;
@@ -185,9 +181,7 @@ export class AggregatorV3Interface extends BaseContract {
       }
     >;
 
-    latestRoundData(
-      overrides?: CallOverrides
-    ): Promise<
+    latestRoundData(overrides?: CallOverrides): Promise<
       [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
         roundId: BigNumber;
         answer: BigNumber;
