+++ autotasks/quickswap-rewards-getter/deploy.sh
