+++ autotasks/quickswap-rewards-getter/src/config.ts
@@ -1,4 +1,4 @@
-import { getDeployedVaultsQuickswap } from "../../../../consts/deployed";
+import { getDeployedVaultsQuickswap } from "../../../consts/deployed";
 
 export const MATIC_CHAIN_ID = 137;
 export const VAULTS = Object.entries(getDeployedVaultsQuickswap()).map(
