+++ autotasks/quickswap-token-depositor/src/config.ts
@@ -1,5 +1,5 @@
-import { NETWORK_ADDRESSES, SECONDS_IN_ONE_DAY } from "../../../../consts";
-import { getDeployedVaultsQuickswap } from "../../../../consts/deployed";
+import { NETWORK_ADDRESSES, SECONDS_IN_ONE_DAY } from "../../../consts";
+import { getDeployedVaultsQuickswap } from "../../../consts/deployed";
 
 export const MATIC_CHAIN_ID = 137;
 export const DEADLINE_SECONDS = SECONDS_IN_ONE_DAY;
