+++ autotasks/quickswap-token-depositor/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "noImplicitAny": true,
     "allowJs": true,
