+++ autotasks/upload-autotasks.ts
@@ -1,5 +1,5 @@
 import { AutotaskClient } from "defender-autotask-client";
-import configEnv from "../../config";
+import configEnv from "../config";
 
 const {
   DEFENDER_TEAM_API_KEY,
