+++ package.json
@@ -15,9 +15,9 @@
     "deploy:dummy-vaults": "hardhat run scripts/deploy-all-dummy-vaults.ts",
     "deploy:dummy-vaults:matic": "npm run deploy:dummy-vaults -- --network matic ",
     "deploy:autotasks": "npm run deploy:autotask:aave-harvester && npm run deploy:autotask:quickswap-rewards-getter && npm run deploy:autotask:quickswap-token-depositor",
-    "deploy:autotask:aave-harvester": "./scripts/defender-autotasks/aave-harvester/deploy.sh",
-    "deploy:autotask:quickswap-rewards-getter": "./scripts/defender-autotasks/quickswap-rewards-getter/deploy.sh",
-    "deploy:autotask:quickswap-token-depositor": "./scripts/defender-autotasks/quickswap-token-depositor/deploy.sh",
+    "deploy:autotask:aave-harvester": "./autotasks/aave-harvester/deploy.sh",
+    "deploy:autotask:quickswap-rewards-getter": "./autotasks/quickswap-rewards-getter/deploy.sh",
+    "deploy:autotask:quickswap-token-depositor": "./autotasks/quickswap-token-depositor/deploy.sh",
     "pretest": "hardhat compile",
     "test": "hardhat test --config test.config.ts",
     "test:types": "tsc --project ./tsconfig.json --noEmit",
